csrdu: align data even if there aren't any colinds
There is an issue when you have both jmp and align set in
csrdu. aligning data always (even if you don't have any
colinds, which means that the size of the unit is 1) makes
decompressing easier.

Also fix a bug in the csdru_test and add a test for the
latter bug.