drle: fix a bug in encoding
Not sure why this used to work or what the fsck I was thinking :-/

EncodeRow() used to pass a col in doEncode(), which (at least now)
looks very bogus. Removed this argument and let doEncode() keep track
of the current column by itself.

Also added a test for the failed case (part of 008.nasasrb.mtx.sorted)