csx: fix a bug in jmp calculation
Encoding ensures that the previous element of the pattern
can be the jmp.

For example:

xs:    [ 10 11 12 13 14 15 ]
delta: [ 10  1  1  1  1  1 ]
drle:  [ (10,1) (1, 5) ]

Hence, we can create a pattern:
 (jmp=10, delta=1, size=5)

The problem is that it is possible for the previous element
to dissapear if it is the member of another pattern.

Applied a quick-n-dirty solution where ->getNextX(x0)  is
called to check for the above.

Added also a test for the bug.