store and use the row start for each spm
row start for each csx thread was not stored and used. This
went unnoticed in the tests, since csx ->SetElems() was
not aware of it either, and each thread end up with a row
jump at the start.

This caused a subtle bug, triggered by a race condition.
At the check for the new row in the multiplication code
all threads besides the first stored the current yr value
to y[0]. If the first thread had allready commited its
yr value, then the value recorded was wrong.

Did some tests, and new scheme seems to work.
More tests needed.