Optimized window sorting.
When extracting windows from the target matrix, only rowptr is copied. The
elems of the window are backed by the the original matrix. Now the window is
more like a partial view of the original matrix.

New method added: SPM::getWindow(), which extracts a window without copying
elems.

To support this feature, changes were made to SPM, SPM::Builder, and also
dynarray, where a new method was added, dynarray_seek(). So you will need this
updated version to sucessfully compile.

Corrected linking in patterns/Makefile.