Generation of statistics can now be achieved by sampling.
Sampling of the matrix is performed on a per-window basis. Whether a window
will be processed or not depends on the sampling probability. All windows have
the same probability to be selected or not (equals to the user-supplied
sampling probability). The user supplies the sampling probability as a real
number in [0,1] through the SAMPLING_PROB environment variable.

Experiments in two sample matrices showed considerable (up to 6x) improvement
in preprocessing times sampling only 5% of the matrix without differences in
the final encoding decision. More experiments are needed, however.

The preprocessing time is about 300 SpMV operations (when non-horizontal
encodings are selected) and is now bound from the encoding phase, which still
sorts the whole matrix.