Changed serial encoding of patterns (EncodeSerial)
It is not necessary to call `genAllStats' when encoding a series of
patterns. However, this is only feasible if we know the deltas to encode. This
is done through a new environment variable, ENCODE_DELTAS, which is in par
with the XFORM_CONF. ENCODE_DELTAS stores a comma-separated list of delta
values in complete correspondance with the pattern types specified in
XFORM_CONF, i.e., first delta value in ENCODE_DELTAS corresponds to the first
pattern type in XFORM_CONF. If ENCODE_DELTAS is set, then the sequence of
pattern types in XFORM_CONF is important, since it specifies the sequence of
encoding. If the ENCODE_DELTAS list has less elements than XFORM_CONF, then
the last value of the list is propagated along.

It is important to note that specifying explicitly a delta, it does not mean
that this delta pattern will be encoded. If it cannot be encoded, then the
method falls back to normal CSR-DU. Only patterns found with the RL encoding
can be included (see note below), which means that setting ENCODE_DELTAS
variable requires prior knowledge of the matrix structure, especially for
block patterns.

Note: A matrix can contain 2x2 blocks, for example, but the RL encoding might
group them greedily in, say, 2x24 blocks. In this case, setting ENCODE_DELTAS
to 2, will not encode any pattern! So this option is mainly useful, if you
know a priori the RL patterns in the matrix, and you don't want to pay again
the cost of generating the statistics.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# Committer: Vasileios Karakasis <bkk@chaos.(none)>
#
# On branch bkk
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   drle.cc
#	modified:   drle.h
#	modified:   spmv.cc
#
# Untracked files:
#   (use "git add <file>..." to include in what will be committed)
#
#	clone_all.out
#	clone_blocks.out
#	clone_horiz.out
#	clone_linear.out
#	csx-results-2010-07-29.tgz
#	csx_llvm_tmpl.llvm.bc
#	gmon.out
#	plots.kkourt/
#	proto/tiling.py
#	results/
#	scirouter
#	sort_windows_old.cpp
#	spmv
#	tags
#	windows-sort
#	../scripts/calc_bits
#	../scripts/mmf_eq
#	../spmv