Preprocessing revisited. Two major changes have been made in the preprocessing step:
(a) Preprocessing is no more probabilistic. We use SAMPLES windows
equally distributed over the whole matrix. For example, if SAMPLES=1
we will take the median matrix window for preprocessing, if SAMPLES=2
we will take two windows, such that they split the matrix into three
(roughly) equal parts. This technique renders preprocessing
deterministic, yielding consistent results in multithreaded
execution. The incosistency of preprocessing results in multithreaded
context was the main reason of dropping the probabilistic approach,
since there was no efficient way to provide a single consistent pseudo
random number sequence across the different threads. As a result of
this change, the SAMPLING_PROB environment variable is also dropped.

(b) We added the option of sampling a specific percentage (portion) of
the input matrix. In this case the window size is determined
automatically from this value and the required number of samples. For
example, if 1% of the input matrix needs to be sampled by a single
sample, we adjust the windows such that each one takes up 1% of the
nonzero elements of the matrix and select the sample according to the
technique described previously. Similarly, we adjust the windows for
different portions and different number of samples by following the
formula:

nr_samples*window_size = percent*nnz/100