Fixed bug with the statistics correction factor. Now we keep the window sizes (in nonzeros) from the beginning and the correction factor is now calculated precisely as following
own_matrix_nr_nzeros / nr_nzeros_of_currert_window

Relying on the value of window->nr_nzeros for the correction factor is
insufficient for matrices with more than one level of encoding, since
this value corresponds to the non-encoded nonzeros of the window. So
using this for calculating the correction factor will lead to
augmented percentage values for patterns in deeper encoding levels.