Full NUMA support for BCSR and VBL.
The NUMA implementation of CSR was also changed to follow the common
pattern.

The NUMA implementation is based on the alloc_interleaved() function
(numa_util.c), which allocates memory with a user-defined interleaving
scheme. For the custom interleaving allocation scheme three parameters
are needed:

(a) the total size (in bytes) of the allocation
(b) the sizes (in bytes) of the individual partitions, and
(c) the list of nodes, where each partition must reside.

The returned allocated array is contiguous in virtual memory (single
call to mmap()), but individual partitions may reside physically on
different memory nodes. This allows a completely transparent use of
the allocated area, yet in a NUMA-aware fashion!

This interleaving scheme renders the conversion of every sparse matrix
storage format rather straightforward, since the partitioning of the
matrix is already known after going multithreaded. So to make the
storage NUMA-aware, we simply re-allocate the data structures using the
required interleaving, copy the elements of the originals to the
NUMA-aware ones, free the originals, and replace them with the new
ones. The change is completely transparent to _multiply() and _size()
functions. Of course, 'destructors' must be customized accordingly
(calls to free_interleaved()).

Finally, we use the numa counterparts of the check/bench_loop()
functions.

Notes:
vector_create_interleaved() must also be adapted to use most general
alloc_interleaved().