New JIT framework.
We abandon the direct construction of the LLVM IR through the LLVM
library interface. Instead, we programmaticaly use the Clang LLVM
front-end framework to generate the LLVM IR code from simple C99 code.

The design is similar to the old JIT and the public interface of CsxJit
remains unchanged. The only "visible" difference is the elimination of
the global calls to initialize the LLVM context and the single
module. Now CsxJit keeps private instances of both to allow for
multi-threaded JIT-ting.

The design of JIT is similar to the old one. We load the initial CSX
SpMV multiplication template (csx_spmv_tmpl.c) which has a number of
special text hooks, which will be filled with the generated C99
code. Responsible for the substitution of the hooks is the TemplateText
class, which behaves similarly to the Python's `string.Template'
class. The generated code is then passed to the ClangCompiler instance
of the CsxJit, which is responsible for setting up the compilation with
the Clang front-end. The actual compilation of the generated source
takes place in the ClangCompiler::Compile() method, where the compiler
is instructed to generate LLVM IR code (EmitLLVMOnlyAction). We also use
some helper functions (jit_util.h) for reading/writing the source
from/to file on disk, creating temporary files, etc.

Finally, the design of the CSX multiplication routine is also
changed. We don't use the switch statement anymore for selecting the
pattern-specific multiplication routines. Instead, we use a
function-table design, which we fill with function pointers to the
pattern-specific multiplication routines. Now the body of the
multiplication is just an offset lookup in the function table based on
the pattern id, i.e., mult_table[patt_id](...).

Software requirements: LLVM 2.9, Clang front-end.