Support for a less aggressive compression scheme.
In this scheme, the initial column index of patterns or delta units is
not stored as a variable-length field keeping the delta distance from
the previous column index. Instead, we store the full starting column
index in a fixed-size 32-bit field. As a result the time-consuming
call to ul_get() is replaced by a simple call to u32_get() to fetch
the next column index.

This fix increases the matrix size up to 2.5%, but it can yield
performance benefits of more than 15% in systems with high memory b/w (e.g.,
Intel Nehalem, NUMA impl., no HT), where the computational part of the
kernel plays a much more significant role.

We also made the following changes:

* Added an interface for enabling/disabling one-dimensional blocks.
* The `numa_lib.sh' is now aware of the CPPFLAGS/LDFLAGS passed to the
  calling Makefiles, in order to correctly find a libnuma instance
  installed in a non-standard location.