New and cleaner implementation of the `fix_interleaving()' function.
New and cleaner implementation of the `fix_interleaving()' function in
`numa_util.c', which is responsible for fixing the user-specified
interleaved allocation according to the system's requirements. The new
implementation proceeds in two distinct phases:

(a) Merging of very small partitions (less than the system's page
    size), in order to meet the system's allocation granularity. The
    partitions are merged from left to right, starting from a small
    partition and continuing until the size of the formed partition
    exceeds the system's page size. The `parts' array is updated
    accordingly; zeros are set in merged partitions. Each newly formed
    partition is assigned at the node with the largest allocation size
    according to the original partition list.

b) Fixing the partition boundaries, so that each partition starts at a
    multiple of the system's page size.

The `alloc_interleaved()' interface changed also slightly: the `nodes'
argument is no more `const int *', since we `fix_interleaving()'
updates it with the nodes of the merged partitions. Thus, the changes
in caller places.

Finally, this implementation also fixes two minor issues of the
previous one:

(a) Fixed memory leak of the old `page_to_nodes' struct. We now use a
    `node_score' struct.

(b) Replaced the use of `numa_num_possible_nodes()' with the correct
    `numa_num_configured_nodes()'. The first is the maximum number of
    nodes supported by the libnuma implementation (namely, 64K), while
    the latter is the actual number of nodes configured by the kernel
    (see man numa(3)).