Revisited allocations and new dynamic array implementation.
All allocations are now handled in a C++ way, through the use of a set
of custom C++ allocators, including a NUMA-aware one (actually a wrapper
of the libnuma library).

The use of the `dynarray' data structure is becoming deprecated and is
gradually replaced by an efficient and flexible C++ implementation
(`DynamicArray' class). The new implementation resembles the STL vector
one, but allows for efficient expansions through the use of a special
set of allocators, supporting a realloc function; `dynamic_array.cc'
contains a small benchmark comparing the different implementations (STL,
C dynarray and C++ DynamicArray). Additionally, support for NUMA-aware
allocations by the DynamicArray is provided transparently through the
use of the corresponding allocator.

The final CSX data structure construction was revisited to make use of
the above new features. The old `csx.h' header is now split in two:

	* `CsxManager.h' -- responsible for the construction of the CSX data
      structures
	* `csx.h' -- definitions of the CSX data structures, also included
	  by the generated SpMV C routines.

The construction of CSX's ctl data structure is now delegated to the
CtlBuilder class.