Added a test sparse matrix for reproducing CSX-sym construction bug.
Added a very sparse symmetric matrix to reproduce the memory corruption
problem generated when running CSX-sym on parabolic_fem. In
parabolic_fem this leads to SIGSEGV, but in this small test matrix, the
bug simply corrupts memory and either leads to a crash later or not at
all. However, running valgrind reveals the invalid writes at
CsxManager::MakeCsx().

The source of this bug seems to be the allocation of the rows_info_ data
structure to be equal to the non zero elements of the matrix
partition. However, there exist symmetric matrices, such as
parabolic_fem, that their non-zero elements are much less than the rows
of a partition. This is happening because the CSX-sym format gathers all
the diagonal elements to another data structure, leaving the CSX
lower_matrix_ partition with a lot of empty rows. This scenario is
reproduced by the added qsymmetric-very-sparse.mtx.sorted' matrix. Run
as follows:

MT_CONF=0 ./spmv -s test-matrices/symmetric-very-sparse.mtx.sorted

or using valgrind (recommended):

MT_CONF=0 valgrind ./spmv -s \
test-matrices/symmetric-very-sparse.mtx.sorted 2> valgrind.out

TODO: Possibly replace `symmetric-very-sparse.mtx.sorted' with a larger
test sparse matrix with the same trait in order to always trigger the
SIGSEGV signal at the point where the overflow occurs.