Store-forwarding fix for IndexType = int for x86_64 arch.
Rewrote the `==' and `!=' SparsePartition<I,V>::iterator operators so as
to enable store forwarding at the processor level. The problem was that
the code produced by GCC for the combined logical AND comparison of the
original operator==() packed two 32-bit integers into a single 64-bit
register and the result was later compared to a 32-bit value. This RAW
dependency cannot be serviced by the store-forwarding mechanism of Intel
processors, due the size mismatch of involved types (see "Intel 64 and
IA-32 Architectures Optimization Reference Manual", Section 3.6.5,
"Store Forwarding"). Due to this fix, the preprocessing phase was
improved by roughly 10% on Ivy-Bridge.

Also removed all setters from the Element class.