Major changes in preprocessing
Elements in a SparsePartition are no longer stored in a Dynamic
Array, but a single std::vector that has a floating size and all
transformations are performed in place. This eliminates redundant
copies of elements and is, therefore, more memory efficient. Also,
the Element class is no longer immutable, but its coordinates
can be modified.

The number of samples is now adjusted to the number of threads
according to the relation: ceil(nr_samples/nr_threads).